requirements.txt
gradio
pypdf
langchain
sentence-transformers
faiss-cpu
transformers